+++ FoodApp/src/App.tsx
@@ -1,4 +1,4 @@
-/*import './App.css'
+import './App.css'
 import { Recipe, Ingredient, Ratings } from './types'
 import CreateRecipePage from './pages/CreateRecipePage';
 import RecipeComponent from './components/DisplayRecipeComponent';
@@ -29,27 +29,5 @@ function App() {
     </>
   );
 };
-
-export default App;
-
-*/
-import { useState } from "react";
-
-import "./App.css";
-import { SearchBar } from "./components/searchBar/SearchBar"
-import { SearchResultsList } from "./components/searchBar/SearchResultsList";
-
-function App() {
-  const [results, setResults] = useState([]);
-
-  return (
-    <div className="App">
-      <div className="search-bar-container">
-        <SearchBar setResults={setResults} />
-        {results && results.length > 0 && <SearchResultsList results={results} />}
-      </div>
-    </div>
-  );
-}
 
 export default App;
