+++ FoodApp/src/components/searchBar/SearchResultsList.css
@@ -8,5 +8,6 @@
     margin-top: 1rem;
     max-height: 300px;
     overflow-y: scroll;
-    
+    color: black;
+
 }
