+++ FoodApp/src/pages/HomePage.tsx
@@ -1,0 +1,19 @@
+import { useState } from "react";
+
+import { SearchBar } from "../components/searchBar/SearchBar";
+import { SearchResultsList } from "../components/searchBar/SearchResultsList";
+
+function App() {
+  const [results, setResults] = useState([]);
+
+  return (
+    <div className="App">
+      <div className="search-bar-container">
+        <SearchBar setResults={setResults} />
+        {results && results.length > 0 && <SearchResultsList results={results} />}
+      </div>
+    </div>
+  );
+}
+
+export default App;
